Add Elm output to git history for now
This should hopefully get the Elm application working in both dev and
production until I can figure out why the assets aren't working in the
production environment when they're included in the .gitignore.

Would love to get this fixed sooner rather than later because the
commits get messy when they include changes to both Elm source code and
the resulting JS output. Should be able to compile assets on deploy (the
application works fine in development).