Convert most of header to Elm SPA
Kind of in between worlds at the moment. I think the long-term idea is
to handle as much of the application as possible on the Elm side. So
it's a good idea to continue building more features into the Elm
application.

But some of this is a bit difficult at the moment since we're using
things like `@current_user` on the Phoenix side to determine whether or
not the user is currently authenticated. And it will take some work to
communicate back and forth from Phoenix to Elm.

The Elm routing seems to be working to pull game data, but that isn't
currently how the nav is set up yet since we're still working with
Phoenix routes.