Change sign up and sign in links back to Phoenix
The long-term plan is to handle all of the UI on the Elm side, but for
now I'm still kind of caught in the middle with some features working on
the Phoenix side.

The Elm authentication pages are now there, but don't actually work yet
because we haven't set up the API for players (only games).

With the current implementation, I can still merge in what I have so far
so I can keep moving. The plan for the time being is to leave some of
the player features on the Phoenix side and we'll build out the games
features on the Elm side with a working API that allows us to build out
all CRUD functionality for games.