Refactor Elm import in app.js
As far as I can tell, the issue is that Heroku deploys try to run the JS
code in app.js before the main.js file is actually compiled.

Hoping that I should be able to just guard against the missing file in
the app.js because it'll exist right after the initial compilation and
deploy finishes.