Always rebuild all the things on deploy
Heroku deploys have been one of the more difficult aspects of getting
things running properly.

Altough it's currently compiling the Elm code properly, it stalls if the
compiled output isn't checked into the repo (because the compilation
happens _after_ the app.js file tries to import the results).

It also looks like there might be an issue with the static assets,
because the user sign out feature works in the local environment but not
in production. I may be missing a mix compile, brunch build, or mix
phoenix.digest command that needs to be run in the production
environment.