Resolve migration issues
I was hoping to adjust the migrations to keep the demo repository clean
and match the book content, but it looks like that's a bad idea since
the production application already has some data and I ought not to
alter the migration history. Hopefully this allows for the deploys to
succeed again.