Disable space bar scrolling globally
Since we're using the space bar as a game start button, it was causing
problems on the game page when there were enough scores being displayed
to cause the page to scroll. It was particularly problematic, because
you'd start the game and then not be able to see it. Hopefully this
should keep the game in view when it starts.