Add conditional check for socket server string
This actually just moves the error occurring from the front-end to the
back-end, but it's good to get an error that I can work towards fixing
instead of ignoring anonymous users.

Previously, anonymous users would get a JS console error since it was
trying to append a non-existent token to the socket server URL.

Although that issue is fixed, it's now going to push errors to the score
channel since it has no player_id to store for anonymous users, and I'll
have to figure out if there's a good way to track scores for different
anonymous users over time.