Display player gameplays in leaderboard list (wip)
Managed to inch closer towards rendering the player scores for
individual players. Ran into a couple of the difficulties of working at
the edge of Phoenix and Elm, because it feels like the gameplays should
be nested inside the player API response, but the easier approach for
now seems to be to decode the gameplays and take care of the filtering
on the Elm side.

I should also probably move the `displayPlayerGameplays` boolean value into the
Player type and not the main Model type, but I wasn't sure how to do that
without interfering with the existing decoder.

Either way this is a little better than the approach of linking to the
player show page on the Phoenix side since it allows for keeping
everything in the Elm application.