Work in progress commit
Can't remember where I left off with this feature. I think I had managed
to get a "jump" feature added, but there were some issues with it. I
think the jump needs to be negative since the SVG coordinates originate
from the top left of the window, and when I left off I think the
character was "jumping" into the ground.