Route to home page after player account creation
This isn't a big change, but it's huge in terms of usability. From
looking at FullStory, it was easy to see that players could get confused
when creating a new account and then just seeing their account
information on the player show page. There's not much to see there, and
clicking "Back" didn't seem intuitive.

This commit should allow players to go straight to the main Elm home
page and see a quick call to action with the "Play Now!" button.

Refs: #25